Refactor persist module structure and rename traits
Moved and renamed files in the persist module for clarity: 'module_ext.rs' is now 'base.rs', and 'collectors.rs' moved up a directory. The 'ModuleExport' trait is renamed to 'ModulePersist' to better reflect its purpose. Updated module imports and documentation accordingly, and removed the now-unnecessary 'export' submodule.
