+++ crates/burn-core/src/persist/collectors.rs
