--- crates/burn-core/src/persist/export/mod.rs
@@ -1,5 +1,0 @@
-mod collectors;
-mod module_ext;
-
-pub use collectors::TensorViewCollector;
-pub use module_ext::ModuleExport;
